Merge Eka's Branch

+++ DTO/FavoriteDTO.cs
@@ -5,13 +5,13 @@ namespace olx_be_api.DTO
     public class CreateFavoriteDTO
     {
         [Required]
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
     }
 
     public class FavoriteResponseDTO
     {
         public int Id { get; set; }
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public DateTime CreatedAt { get; set; }
     }
